+++ graphene-writer/src/main/kotlin/com/graphene/writer/config/GrapheneWriterProperty.kt
@@ -7,7 +7,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties
  * @author Andrei Ivanov
  */
 @ConfigurationProperties(prefix = "graphene.writer")
-class GrapheneWriterConfiguration(
+class GrapheneWriterProperty(
   var cassandraDataStoreHandlerProperty: CassandraDataStoreHandlerProperty,
   var statsProperty: StatsProperty
 ) {