+++ graphene-writer/src/main/kotlin/com/graphene/writer/input/graphite/CarbonServerHandler.kt
@@ -3,15 +3,19 @@ package com.graphene.writer.input.graphite
 import com.google.common.base.CharMatcher
 import com.graphene.writer.config.Rollup
 import com.graphene.writer.domain.Metric
-import com.graphene.writer.input.graphite.property.CarbonProperty
+import com.graphene.writer.input.graphite.property.InputGraphiteCarbonProperty
 import com.graphene.writer.processor.GrapheneProcessor
 import io.netty.buffer.ByteBuf
 import io.netty.channel.ChannelHandler
 import io.netty.channel.ChannelHandlerContext
 import io.netty.channel.ChannelInboundHandlerAdapter
 import io.netty.util.CharsetUtil
+import java.util.Objects
 import javax.annotation.PostConstruct
+import javax.annotation.PreDestroy
 import org.apache.logging.log4j.LogManager
+import org.jmxtrans.embedded.QueryResult
+import org.jmxtrans.embedded.output.GraphiteWriter
 import org.springframework.stereotype.Component
 
 /**
@@ -20,19 +24,27 @@ import org.springframework.stereotype.Component
 @Component
 @ChannelHandler.Sharable
 class CarbonServerHandler(
-  private val carbonProperty: CarbonProperty,
+  private val inputGraphiteCarbonProperty: InputGraphiteCarbonProperty,
   private val grapheneProcessor: GrapheneProcessor
 ) : ChannelInboundHandlerAdapter() {
 
   private val logger = LogManager.getLogger(CarbonServerHandler::class.java)
 
   private lateinit var rollup: Rollup
-  private lateinit var graphiteCodec: GraphiteMetricConverter
+  private lateinit var graphiteConverter: GraphiteMetricConverter
+  private lateinit var graphiteWriter: GraphiteWriter
 
   @PostConstruct
   fun init() {
-    this.rollup = carbonProperty.baseRollup!!
-    this.graphiteCodec = GraphiteMetricConverter()
+    this.rollup = inputGraphiteCarbonProperty.baseRollup!!
+    this.graphiteConverter = GraphiteMetricConverter()
+    if (Objects.nonNull(inputGraphiteCarbonProperty.route)) {
+      this.graphiteWriter = GraphiteWriter()
+      this.graphiteWriter.settings["host"] = inputGraphiteCarbonProperty.route!!.host
+      this.graphiteWriter.settings["port"] = inputGraphiteCarbonProperty.route!!.port
+      this.graphiteWriter.settings["namePrefix"] = IGNORE
+      this.graphiteWriter.start()
+    }
   }
 
   @Throws(Exception::class)
@@ -46,7 +58,14 @@ class CarbonServerHandler(
       }
 
       if (CharMatcher.ascii().matchesAllOf(metric.path) && CharMatcher.ascii().matchesAllOf(metric.tenant)) {
-        grapheneProcessor.process(graphiteCodec.convert(GraphiteMetric(metric.path, metric.value, normalizeTimestamp(metric.timestamp))))
+        val grapheneMetrics = graphiteConverter.convert(GraphiteMetric(metric.path, metric.value, normalizeTimestamp(metric.timestamp)))
+        for (grapheneMetric in grapheneMetrics) {
+          grapheneProcessor.process(grapheneMetric)
+
+          if (Objects.nonNull(inputGraphiteCarbonProperty.route)) {
+            graphiteWriter.write(listOf(QueryResult(metric.path, metric.value, metric.timestamp * 1_000L)))
+          }
+        }
       } else {
         logger.warn("Non ASCII characters received, discarding: $metric")
       }
@@ -59,5 +78,16 @@ class CarbonServerHandler(
 
   fun normalizeTimestamp(timestamp: Long): Long {
     return timestamp / rollup.rollup * rollup.rollup
+  }
+
+  @PreDestroy
+  fun destroy() {
+    if (Objects.nonNull(inputGraphiteCarbonProperty.route)) {
+      graphiteWriter.stop()
+    }
+  }
+
+  companion object {
+    const val IGNORE = ""
   }
 }