+++ graphene-writer/src/test/kotlin/com/graphene/writer/input/GrapheneMetricTest.kt
@@ -1,6 +1,7 @@
 package com.graphene.writer.input
 
 import java.util.Collections
+import java.util.TreeMap
 import kotlin.test.assertEquals
 import org.joda.time.DateTimeUtils
 import org.junit.jupiter.api.Test
@@ -11,20 +12,22 @@ internal class GrapheneMetricTest {
   internal fun `should return graphite key joined with dot ordered by numeric`() {
     // given
     val grapheneMetric = GrapheneMetric(
+      Source.GRAPHITE,
+      "a.b.c",
       Collections.emptyMap(),
-      mapOf(
+      TreeMap(mapOf(
         Pair("0", "a"),
         Pair("1", "b"),
-        Pair("2", "c")
+        Pair("2", "c"))
       ),
       1.0,
       DateTimeUtils.currentTimeMillis()
     )
 
     // when
-    val graphiteKey = grapheneMetric.getGraphiteKey()
+    val id = grapheneMetric.id
 
     // then
-    assertEquals(graphiteKey, "a.b.c")
+    assertEquals(id, "a.b.c")
   }
 }